BobmooCrawlling/requirements.txt
playwright==1.40.0
google-generativeai==0.3.2
beautifulsoup4==4.12.2
python-dotenv==1.0.0
PyPDF2==3.0.1
requests==2.31.0
